refactor: 정적 팩토리 메서드로 리팩토링해서 객체 책임 분리
