+++ src/main/java/webserver/domain/RequestLine.java
@@ -2,6 +2,7 @@ package webserver.domain;
 
 public class RequestLine {
 
+    private static final String REQUEST_LINE_DELIMITER = " ";
     public static final int METHOD_INDEX = 0;
     public static final int URL_INDEX = 1;
     public static final int PROTOCOL_INDEX = 2;
@@ -14,6 +15,16 @@ public class RequestLine {
         this.method = method;
         this.url = url;
         this.protocolVersion = protocolVersion;
+    }
+
+    public static RequestLine from(final String startLine){
+        String[] httpElement = startLine.split(REQUEST_LINE_DELIMITER);
+
+        HttpMethod httpMethod = HttpMethod.from(httpElement[RequestLine.METHOD_INDEX]);
+        Url url = Url.from(httpElement[RequestLine.URL_INDEX]);
+        ProtocolVersion protocol = ProtocolVersion.from(httpElement[RequestLine.PROTOCOL_INDEX]);
+
+        return new RequestLine(httpMethod, url, protocol);
     }
 
     public HttpMethod getMethod() {
