+++ src/main/java/webserver/domain/Url.java
@@ -1,6 +1,7 @@
 package webserver.domain;
 
 public class Url {
+    private static final String URL_DELIMITER = "\\?";
     public static final int PATH_INDEX = 0;
     public static final int QUERY_PARAMETER_INDEX = 1;
 
@@ -10,6 +11,15 @@ public class Url {
     public Url(String path, QueryParameter queryParameter) {
         this.path = path;
         this.queryParameter = queryParameter;
+    }
+
+    public static Url from(final String url) {
+        String[] splitUrl = url.split(URL_DELIMITER);
+
+        return new Url(
+                splitUrl[Url.PATH_INDEX],
+                QueryParameter.from(splitUrl)
+        );
     }
 
     public String getPath() {
